refactor(config): extract typed_value coercion as shared helper
The gateway's PATCH/PUT handlers enforced PropKind validation on
incoming JSON values; the CLI's `config patch` flow used a strictly
weaker coercion that silently round-tripped non-array JSON through
set_prop's string parser. Two contracts for the same data path.

Move both into `zeroclaw_config::typed_value::coerce_for_set_prop`
and have CLI + gateway delegate. CLI now rejects the same shape
mismatches HTTP rejects. The CLI helper resolves the field's
PropKind via `Config::prop_fields()` against the live config so the
caller doesn't need to plumb it through.

Eight unit tests in the new module cover null-resets, array element
type checking with offending-index reporting, empty-array-as-valid,
bool/integer/float strictness, legacy bool-string accept, and
unknown-kind fallback. Three duplicate gateway tests removed (they
covered the same surface).

Refs #6175. Adversarial review pass 1, finding P1.2.
